# docs/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgriScope Thailand API",
    "version": "0.1.0",
    "description": "Contract-first API for multi-tenant agricultural satellite analysis."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "accessCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agriscope_access"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "access token"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agriscope_refresh"
      },
      "csrfCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agriscope_csrf"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication required.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "CSRF validation or role authorization failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Scoped resource not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Request validation failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "RateLimitedError": {
        "description": "Rate limit exceeded.",
        "headers": {
          "Retry-After": {
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id",
              "details"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "request_id": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "example": {
          "error": {
            "code": "validation_error",
            "message": "Request validation failed",
            "request_id": "018f1b46-0637-7000-9000-000000000001",
            "details": {}
          }
        }
      },
...
      "Organization": {
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "logo_url": {
            "type": "string",
            "nullable": true
          },
          "plan_code": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended",
              "deleted"
            ]
          }
        },
        "example": {
          "id": "018f1b46-0637-7000-9000-000000000010",
          "name": "Example Farm Co",
          "slug": "example-farm-co-a1b2c3d4",
          "plan_code": "starter",
          "status": "active"
        }
      },
...
      "Member": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "role",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": [
              "organization_owner",
              "organization_admin",
              "agronomist",
              "field_manager",
              "viewer"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled",
              "invited"
            ]
          }
        }
      },
      "OrganizationCreateRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "example": {
          "name": "North Field Cooperative"
        }
      },
...
      "SatelliteAcquisition": {
        "type": "object",
        "required": [
          "provider",
          "collection",
          "item_id",
          "acquired_at"
        ],
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "cdse_stac"
            ]
          },
          "collection": {
            "type": "string",
            "enum": [
              "sentinel-2-l2a"
            ]
          },
          "item_id": {
            "type": "string"
          },
          "acquired_at": {
            "type": "string",
            "format": "date-time"
          },
          "cloud_cover_percent": {
            "type": "number",
            "nullable": true
          }
        },
        "example": {
          "provider": "cdse_stac",
          "collection": "sentinel-2-l2a",
          "item_id": "S2B_MSIL2A_20260731T035529_N0511_R004_T47PPR_20260731T072142",
          "acquired_at": "2026-07-31T03:55:29Z",
          "cloud_cover_percent": 12.4
        }
      },
      "SatelliteAvailableResponse": {
        "type": "object",
        "required": [
          "field_id",
          "status",
          "acquisition",
          "searched_at",
          "message_th"
        ],
        "properties": {
          "field_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "const": "available"
          },
          "acquisition": {
            "$ref": "#/components/schemas/SatelliteAcquisition"
          },
          "searched_at": {
            "type": "string",
            "format": "date-time"
          },
          "message_th": {
            "type": "string"
          }
        }
      },
      "SatelliteNotSearchedResponse": {
        "type": "object",
        "required": [
          "field_id",
          "status",
          "acquisition",
          "searched_at",
          "message_th"
        ],
        "properties": {
          "field_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "const": "not_searched"
          },
          "acquisition": {
            "type": "null"
          },
          "searched_at": {
            "type": "null"
          },
          "message_th": {
            "type": "string"
          }
        }
      },
      "SatelliteEmptySearchResponse": {
        "type": "object",
        "required": [
          "field_id",
          "status",
          "acquisition",
          "searched_at",
          "message_th"
        ],
        "properties": {
          "field_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "no_data",
              "temporarily_unavailable"
            ]
          },
          "acquisition": {
            "type": "null"
          },
          "searched_at": {
            "type": "string",
            "format": "date-time"
          },
          "message_th": {
            "type": "string"
          }
        }
      },
      "SatelliteLatestResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/SatelliteAvailableResponse"
          },
          {
            "$ref": "#/components/schemas/SatelliteNotSearchedResponse"
          },
          {
            "$ref": "#/components/schemas/SatelliteEmptySearchResponse"
          }
        ],
        "discriminator": {
          "propertyName": "status",
          "mapping": {
            "available": "#/components/schemas/SatelliteAvailableResponse",
            "not_searched": "#/components/schemas/SatelliteNotSearchedResponse",
            "no_data": "#/components/schemas/SatelliteEmptySearchResponse",
            "temporarily_unavailable": "#/components/schemas/SatelliteEmptySearchResponse"
          }
        }
      },
      "SatelliteSearchResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/SatelliteAvailableResponse"
          },
          {
            "$ref": "#/components/schemas/SatelliteEmptySearchResponse"
          }
        ],
        "discriminator": {
          "propertyName": "status",
          "mapping": {
            "available": "#/components/schemas/SatelliteAvailableResponse",
            "no_data": "#/components/schemas/SatelliteEmptySearchResponse",
            "temporarily_unavailable": "#/components/schemas/SatelliteEmptySearchResponse"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/organizations": {
      "post": {
        "summary": "Create an organization for the current user.",
        "security": [
          {
            "accessCookie": [],
            "csrfCookie": [],
            "csrfHeader": []
          },
          {
            "bearerAuth": []
          }
        ],
        "x-required-role": "authenticated user.",
        "x-organization-scope": "creates organization and owner membership.",
        "x-validation": "name required, slug generated collision-safe.",
        "x-rate-limit": "trusted-client-IP 300 and authenticated subject 60 per configured window.",
        "x-idempotency": "not idempotent.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationCreateRequest"
              }
            }
          }
        },
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitedError"
          },
          "201": {
            "description": "Created organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List organizations visible to current user.",
        "security": [
          {
            "accessCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "x-required-role": "active membership.",
        "x-organization-scope": "scoped by active memberships for current user.",
        "x-validation": "disabled memberships excluded.",
        "x-idempotency": "safe read-only operation.",
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "200": {
            "description": "Organizations.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Organization"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{organization_id}": {
      "get": {
        "summary": "Get organization if current user has active membership.",
        "security": [
          {
            "accessCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "x-required-role": "viewer",
        "x-organization-scope": "query scopes by organization id and active membership.",
        "x-validation": "invalid UUID rejected; foreign organization returns 404.",
        "x-idempotency": "safe read-only operation.",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "200": {
            "description": "Organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          },
          "404": {
            "description": "Not found or not visible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{organization_id}/members": {
      "get": {
        "summary": "List organization members for active members.",
        "security": [
          {
            "accessCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "x-required-role": "viewer",
        "x-organization-scope": "organization id and current active membership required.",
        "x-validation": "disabled membership denied; foreign organization returns 404.",
        "x-idempotency": "safe read-only operation.",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "200": {
            "description": "Members.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Member"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/fields/{field_id}/satellite/search-latest": {
      "post": {
        "summary": "Search CDSE STAC for latest Sentinel-2 L2A acquisition intersecting a saved field.",
        "security": [
          {
            "accessCookie": [],
            "csrfCookie": [],
            "csrfHeader": []
          },
          {
            "bearerAuth": []
          }
        ],
        "x-required-role": "viewer",
        "x-organization-scope": "field lookup joins active membership before provider request.",
        "x-validation": "persisted field geometry only; browser geometry is ignored; deleted or foreign fields return 404.",
        "x-timeout": "provider timeout returns status temporarily_unavailable without stack traces.",
        "x-rate-limit": "authenticated subject 20 and subject plus HMAC(field ID) 10 per configured window.",
        "x-idempotency": "repeated searches reuse field/provider/item acquisition records.",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitedError"
          },
          "200": {
            "description": "Latest acquisition search result, no-data result, or temporary provider-unavailable result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SatelliteSearchResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Field not found or not visible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fields/{field_id}/satellite/latest": {
      "get": {
        "summary": "Return persisted latest Sentinel-2 L2A acquisition for a saved field.",
        "security": [
          {
            "accessCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "x-required-role": "viewer",
        "x-organization-scope": "field lookup joins active membership before acquisition read.",
        "x-validation": "deleted or foreign fields return 404.",
        "x-timeout": "no provider request is made.",
        "x-idempotency": "safe read-only operation.",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "200": {
            "description": "Persisted available acquisition or not_searched with null acquisition and searched_at when no result is stored.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SatelliteLatestResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Field not found or not visible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}